# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sportswear E-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filtering
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: bestSeller
          in: query
          schema:
            type: ["boolean", "null"]
        - name: newArrival
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Add a review for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Categories
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        sku:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - comparePrice
        - imageUrl
        - images
        - categoryId
        - categoryName
        - sizes
        - colors
        - stock
        - sku
        - isFeatured
        - isBestSeller
        - isNewArrival
        - averageRating
        - reviewCount
        - createdAt
        - updatedAt
    ProductList:
      type: array
      items:
        $ref: "#/components/schemas/Product"
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        sku:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
      required:
        - name
        - description
        - price
        - imageUrl
        - stock
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        sku:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
...
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productId
        - customerName
        - rating
        - comment
        - createdAt
    ReviewList:
      type: array
      items:
        $ref: "#/components/schemas/Review"
    CreateReviewBody:
      type: object
      properties:
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required:
        - customerName
        - rating
...